[1 BUG FIXED]
The BUG related to diagonal attack not being handled properly, because the
mask calculated for the diag attacks being wrong has been fixed.

The mask was becoming wrong because, as the base mask was getting shifted
to the left or right, the squares of the diag which was getting off the
board was in reality getting wrapped around to the next or previous rank,
because the board was being represented has a bitboard.

* Added a debug command thro a new debugcmd.c file
  * Inturn a subcommand diagattack added to check the diagattack mask
    was getting generated properly. The logic for printing the diagattack
    debug message is protected using the #define DEBUG_DIAGATTACK

* Fixed the diagattack mask logic as mentioned previously.