[GOOD - debug log cleanup - AND SPEED Increase]
* Added Make variables for some of the common options
  The Debug version has all the key debug logs enabled WHILE
  The Non Debug version has most of the debug logs disabled.

* size of fixed buffer used by cb_2longnot increased to standard MOVES_BUFSIZE

* Fenced adding move number to sMoves in struct cb. This is disabled by
  default. Also renamed the #define fencing the NextBestMoves move number
  adding logic.

* Removed sending info UCI messages of leaf moves during depth search.
  In turn replaced it with Fenced debug logging which is disabled for the
  normal non debug version of the executables.

* Updated/Cleaned up the debug logs related to positions which have no valid
  moves for the sideToMove.

* Updated the logic which sends uci info messages during depth search unwinding
  such that it sends them only for the last 2 unwinding depths.
  Also fenced the corresponding debug logs for the other depths of the search.

* Updated the info being sent to fLogM

* Took care of flushing the file buffers at end of process_go

* In general for the normal non debug version the debug logs have been reduced.
  Instead the debug version has these logs enabled.

  Also the unwind selection debugging logs have been removed from normal
  version.

NOTE: The above changes (mainly reducing the amount of uci info messages being
sent -because it was also being flushed immidiately) has speeded up the engine
working by almost 4 to 6 times on average, and lot more faster in some cases.

TOFIX: Also found that when adding position eval to HashTable, I am not taking
care of also storing the depth at which it was evaluated and stored. This
inturn leads to moves being dropped too early if the position repeats early
in the depth search for a different move combination. Where has it should have
actually been searched further as it would have led to a more appropriate
value for the position. THIS was easily found when the program depth was
changed from 5 to 6, as this gave lot of oppurtunity for the same positions to
repeat at different depths.