[ABPruning - cleanup/cornercases - TOCHECKLATER] Found that as I moved from search depth of 6 or 7 to around 9 to 10. Even thou still much faster than the older non ABPruning version. Still it was taking some relatively more time than I had expected. So did the following
1. Previously I was short circuiting a search only if its value was relatively
Worse than what was previously found. But Now I short circuit it even in cases
where they may be same.

NOTE:TOTHINK: My original idea was to use additional statistics to decide
which among the many equal moves to select (based on which leads to a better
position at the end). This can be achieved using

1.1. Having a better eval function which evaluates better, so less chances of
two positions evaluating to same value.
OR
1.2. As previously thought have a summation of evals of all the moves for a
given position and then use that to compare between equal positions to see
which is better.

2. The previous aging logic which was added to find move lists which are short
to get to a given eval rather than a equivalent longer move list has been
fenced out for ABPruning. HAVE TO TEST THE IMPACT OF THIS CHANGE LATTER MORE
THOROUGHLY.