[General - GameState - Start/Mid/End]
* Added logic to evalpw_mat to identify whether the game is at the start, mid
  or the end phase based on the amount of material still remaining in the board.

  * The same is used to set a new state variable in the chess board struct.

  * The total material count is also returned from this function.

* This game state is used to see to it that Pawn movement is given a slight
  additional weightage during Start and End phases.

  * The logic in evalpw_pos made bit more decoupled from one another now.

NOTE:TODO: While trying to reenable the Knight position eval logic, it was
found that pawn enpassant was occuring few times, as this is not currently
handled by my engine, those games invariably lead to a loss, as my engine
assumes a pawn on board, when it is no longer there. This is as expected,
however to advance the engine further, this missing move requires to be
implemented first, before adding any other features.

NOTE: Have added two different thresholds for MIDGAME, the commented one
may be better in some cases. Have to finetune this later.