[SpecialMoves - Understand Pawn EnPassant]
* Added flags to chess board structure to track if Enpassant is possible
  in the next move or not.

* The low level move helper sets the Enpassant flag for the opponent if the
  current side has made a double square move wrt the pawn.

* If the flag is set such that the current side can do a enpassant move
  with its pawn, then it cross checks if the current move is a enpassant move
  and then if it is, then it does the needful.

* If the Enpassant move was possible and the opponent didn't execute the same
  then the Enpassant flag is cleared.

  NOTE:FIXME:TODO: From a optimisation perspective maybe I can directly check
  for zero. Have to check the assembly code generated in both cases and see
  which is better later.

NOTE: This logic takes care of identifying a enpassant move sent by the UCI
gui and currently the Engine cann't do a enpassant move of its own. However
the logic for setting the flag as well as executing the actual move is fully
implemented in the lowlevel move helper, SO it will be very easy in future
to add support for the engine to execute its own enpassant move. TODO: Basicaly
the Pawn Move generation logic should check if a enpassant move is possible
and if it is possible, then add a enpassant move to the list of moves.