QuickFindBest P1... forgotten 20131224 in 20181012
I am doing this commit in 20181012. But I had last looked at and
worked on my chess logic in 20131224...

So I have forgotten the original idea which I had in mind when I
was trying to do this drastic change to the logic.

Also I dont remember the logic fully now. HOWEVER WHAT I SEEM TOBE
doing here (based on a very casual quick glance) is that

TAKE1
----------

Instead of sending a static eval value on reaching max depth or
if king is under attack,

Now I set a hint to tell that I should be doing a Fast/Quick find
best logic, when ever the curDepth is equal to or greater than
maxDepth.

In turn in the qfindbest for king under check, I send the static
eval value.

However in case the king is not under check what I have simplified
in the qfindbest when compared to findbest, I require to crosscheck

TAKE2
--------

Also I have removed the following from qfindbest

* multiple threads based logic

* ABSHORTKILLED related logic

* if King enters check then return KingEntersCheckEval logic

* DEBUG_UNWIND related logic

* MOVELIST_ADDMOVENUM_NBMOVES related logic

* USE_MOVELISTEVALAGING related logic

* Sending of info messages and logging of debug info at the end

* USE_HASHTABLE related logic at the end

Instead of sending WHITE/BLACKSTUCK_GOODFORBLACK/WHITE value now
send the passed bestW or bestB respectively.