Move Generation sequencing cleanup for speed
Have changed the flow of code in the available moves generation logics
for bishop, rook and queen for the case of when other things block
them from going the full length of a file or a rank or a diagonal.

This should hopefully help speed up the move generation a bit.

Currently as I am not profiling the code properly, I cant say there
as been a definite improvement. But technically there should be,
TOCHECK_LATER.