Cleanup and Simplification and More Threads
a) Have removed unused conversion of eval value from whites perspective
to orig-side-to-move perspective

b) Have increased the number of threads to 8 yesterday night, to test
on my new i7-8750H laptop

c) Instead of calculating the static eval value always in findbest,
do it only if

* either the king is killed now (from the last move which lead to this
position being looked at)

* or if static eval only hint has been set.

c.1) Initially I had made it such that king is killed check will be
valid only in the next move, after a previous move has already killed
the king. However as was noticed ages back, when I had tried a similar
thing, it lead to the engine losing a game, because the right move
was not getting selected, if the king is killed (TODO: MAYBE bcas,
the king killed flag is not propogated properly across moves as one
goes down the moves tree, so that info was lost oddly, seems bit odd,
but have to check this possibility/bug ---- RATHER in this case
currently as the static eval would have only occured in qfindbest,
and not the next move after king is killed, by then other entities
in the board might have got killed and or moved positions, thus the
static value identified in qfindbest would have been wrong, leading
to wrong move being selected).

However for now I have added a new cb_check_kingkilled which gets
called always in findbest. This ensures that such a situation is
identified in findbest itself immidiately, equivalent to the old logic
with the more processing costly static eval logic (from king killed
perspective).

At same time, the other overheads wrt static eval is now avoided as
one is walking thro the tree towards the leaf (decided by depth
allowed).

With this now I (rather my chess engine) won the game against
stockfish 8, which it had lost previously. Here I restarted that
lost game from 18th move or so, which was somewhere in the middle
of the total moves in the game.

d) As mentioned in NOTEs previously, the engine is currently (over
the last 2 days or so) running with only the material eval enabled,
as the pawn position based eval which was there previously was
disabled few days back(after disabling position eval, I think I
have tested for 2 or 3 days, with gaps in between).